OCR/requirements.txt
streamlit
tensorflow
pytesseract
pillow
